# speed up `.config_key_from_metrics()` helper