--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -35,7 +35,7 @@ Imports:
     tibble (>= 3.1.0),
     tidyr (>= 1.2.0),
     tidyselect (>= 1.1.2),
-    vctrs (>= 0.4.1),
+    vctrs (>= 0.6.1),
     withr,
     workflows (>= 1.0.0),
     yardstick (>= 1.0.0)